refactor(dashboards-v2): move toPostableTags to utils next to its inverse
